chore: move label migrators and helpers
