+++ internal/provider/resource_docker_container_migrators.go
@@ -906,3 +906,41 @@ func updateV0ToV1PortsOrder(is *terraform.InstanceState, meta interface{}) error
 
 	return nil
 }
+
+func replaceLabelsMapFieldWithSetField(rawState map[string]interface{}) map[string]interface{} {
+	labelMapIFace := rawState["labels"]
+	if labelMapIFace != nil {
+		labelMap := labelMapIFace.(map[string]interface{})
+		rawState["labels"] = mapStringInterfaceToLabelList(labelMap)
+	} else {
+		rawState["labels"] = []interface{}{}
+	}
+
+	return rawState
+}
+
+func migrateContainerLabels(rawState map[string]interface{}) map[string]interface{} {
+	replaceLabelsMapFieldWithSetField(rawState)
+
+	m, ok := rawState["mounts"]
+	if !ok || m == nil {
+		// https://github.com/terraform-providers/terraform-provider-docker/issues/264
+		rawState["mounts"] = []interface{}{}
+		return rawState
+	}
+
+	mounts := m.([]interface{})
+	newMounts := make([]interface{}, len(mounts))
+	for i, mountI := range mounts {
+		mount := mountI.(map[string]interface{})
+		volumeOptionsList := mount["volume_options"].([]interface{})
+
+		if len(volumeOptionsList) != 0 {
+			replaceLabelsMapFieldWithSetField(volumeOptionsList[0].(map[string]interface{}))
+		}
+		newMounts[i] = mount
+	}
+	rawState["mounts"] = newMounts
+
+	return rawState
+}
